requirements.txt
fastapi                    
uvicorn[standard]           
sqlalchemy                  
pymysql                     
python-dotenv               
pydantic                    
python-jose[cryptography]   